Refactor project startup
Change JWT configuration

Configuration clean-up

Clean up project startup code
