+++ EPlast/EPlast.WebApi/appsettings.json
@@ -1,8 +1,4 @@
 {
-  "Admin": {
-    "Email": "",
-    "Password": ""
-  },
   "Logging": {
     "LogLevel": {
       "Default": "Information",
@@ -11,51 +7,36 @@
     }
   },
   "AllowedHosts": "*",
-  "ConnectionStrings": {
-    "EPlastDBConnection": "",
-    "Redis": "10.0.95.244:6379"
+  "BlobContainerNames": {
+    "AboutBaseImages": "aboutbaseimages",
+    "UserImages": "userimages",
+    "DecisionsFiles": "decisionfiles",
+    "MethodicDocumentsFiles": "methodicdocumentsfiles",
+    "BlankImages": "blankimages",
+    "AchievementContainer": "achievementcontainer",
+    "ExtractFromUPUContainer": "extractfromupucontainer",
+    "CityImages": "cityimages",
+    "CityFiles": "cityfiles",
+    "RegionImages": "regionimages",
+    "RegionFiles": "regionfiles",
+    "ClubImages": "clubimages",
+    "ClubFiles": "clubfiles",
+    "EventsImages": "eventsimages",
+    "GoverningBodiesImages": "governingbodiesimages",
+    "GoverningBodyFiles": "governingbodyfiles",
+    "GoverningBodySectorsImages": "governingbodysectorsimages",
+    "GoverningBodySectorsFiles": "governingbodysectorsfiles"
   },
-  "StorageConnectionString": "",
-  "AboutBaseImages": "aboutbaseimages",
-  "UserImages": "userimages",
-  "DecisionsFiles": "decisionfiles",
-  "MethodicDocumentsFiles": "methodicdocumentsfiles",
-  "BlankImages": "blankimages",
-  "AchievementContainer": "achievementcontainer",
-  "ExtractFromUPUContainer": "extractfromupucontainer",
-  "CityImages": "cityimages",
-  "CityFiles": "cityfiles",
-  "RegionImages": "regionimages",
-  "RegionFiles": "regionfiles",
-  "ClubImages": "clubimages",
-  "ClubFiles": "clubfiles",
-  "EventsImages": "eventsimages",
-  "GoverningBodiesImages": "governingbodiesimages",
-  "GoverningBodyFiles": "governingbodyfiles",
-  "GoverningBodySectorsImages": "governingbodysectorsimages",
-  "GoverningBodySectorsFiles": "governingbodysectorsfiles",
   "EmailServiceSettings": {
     "SMTPServer": "smtp.ukr.net",
     "Port": 465
   },
   "GoogleAuthentication": {
-    "GoogleClientId": "",
-    "GoogleClientSecret": "",
     "GoogleApiTokenInfoUrl": "https://www.googleapis.com/oauth2/v3/tokeninfo?id_token={0}"
-  },
-  "FacebookAuthentication": {
-    "FacebookAppId": "",
-    "FacebookAppSecret": ""
   },
   "URLs": {
     "SignIn": "https://eplast.westeurope.cloudapp.azure.com/signin/",
     "Cities": "https://eplast.westeurope.cloudapp.azure.com/cities/",
     "UserPage": "https://eplast.westeurope.cloudapp.azure.com/userpage/main/"
-  },
-  "Jwt": {
-    "Key": "2af4ff57-4ca0-4b3a-804b-178ad27aaf88",
-    "Issuer": "https://localhost:44350/",
-    "Audience": "https://localhost:3000/",
-    "Time": 120
   }
 }
